moved styles files from their own folder
