+++ src/components/AddMovieButton/styles.scss
@@ -1,4 +1,4 @@
-@import "src/theme/constants";
+@import "../../theme/constants";
 
 .add-movie-wrapper {
   max-width: 100%;
