+++ src/components/FunctionalComponent/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 
-import './styles/styles.scss'
+import './styles.scss'
 
 const FunctionalComponent = () => {
   const [counter, setCounter] = useState(0)
