+++ src/components/FunctionalComponent/styles.scss
