+++ src/components/GenreMenuBar/styles.scss
@@ -1,4 +1,4 @@
-@import "./src/theme/constants";
+@import "../../theme/constants";
 
 .genre-menu-bar {
   position: relative;
