+++ src/components/Header/styles.scss
@@ -2,7 +2,7 @@
   height: 50vh;
   max-height: 396px;
   width: 100%;
-  background-image: url('../../../assets/images/Bitmap.png');
+  background-image: url('../../assets/images/Bitmap.png');
   background-repeat: no-repeat;
   background-size: 100%;
   position: relative;
