+++ src/components/SearchBar/styles.scss
@@ -1,4 +1,4 @@
-@import "./src/theme/constants.sass";
+@import "../../theme/constants";
 
 .search-bar-wrapper {
     margin-left: 10vw;
