+++ src/pages/MainPage/styles.scss
